requirements.txt
azure-common==1.1.28
azure-core==1.28.0
azure-identity==1.13.0
azure-mgmt-core==1.4.0
azure-mgmt-resource==23.0.1